Added implementation of object factory to support Java Immutables
This CR adds support for creating instances of an interface that have a
concrete class on the class-path. This supports libraries that generate
POJOs from interfaces.

Also, the interface proxy implementation of the object factory does not
handle parameterized types (i.e. generics). This is now supported, for
concrete classes that support generics

cr https://code.amazon.com/reviews/CR-15845577